Gabriel

+++ Backend/src/services/reunion.service.js
@@ -1,11 +1,12 @@
 import { AppDataSource } from "../config/configDb.js";
+import { transporter } from "./correo.service.js";
 
 export async function createReunionService(datos) {
   try {
-    const { asunto, fechaInicio } = datos;
+    const { asunto, fechaInicio, lugar } = datos;
 
-    if (!asunto || !fechaInicio ) {
-      return [null, "Faltan campos obligatorios"];
+    if (!asunto || !fechaInicio || !lugar) {
+      return ["Faltan campos obligatorios"];
     }
 
 
@@ -13,42 +14,68 @@ export async function createReunionService(datos) {
 
     const existeReunion = await reunionRepo.findOne({ where: { fechaInicio } });
     if (existeReunion) {
-      return [null, "Ya existe una reunión programada para esta fecha"];
+      return ["Ya existe una reunión programada para esta fecha"];
     }
 
     const hoy = new Date();
     const fecha = new Date(fechaInicio);
 
     if (fecha < hoy) {
-      return [null, "La fecha de inicio no puede ser anterior a hoy"];
+      return [ "La fecha de inicio no puede ser anterior a hoy"];
     }
 
     const nuevaReunion = reunionRepo.create({
       asunto,
       fechaInicio,
+      lugar,
       estado: "programada",
     });
 
     await reunionRepo.save(nuevaReunion);
 
-    return [nuevaReunion, null];
+    const usuarioRepo = AppDataSource.getRepository("Usuario");
+
+    const usuarios = await usuarioRepo.find();
+
+    for (const usuario of usuarios) {
+      try {
+        await transporter.sendMail({
+          from: '"Junta de Vecinos" <gabriels.guzmanc@gmail.com>', 
+          to: usuario.email,  
+          subject: `Reunión Programada Fecha: ${fechaInicio}`,
+          text: `Estimados vecinos,\n\nSe ha programado una reunion para la fecha ${fechaInicio} con motivo ${asunto} \n\n Esperamos su asistencia\n\nSaludos cordiales,\nJunta de Vecinos Parque Ecuador`,
+        });
+        console.log("Correo enviado correctamente");
+      } catch (error) {
+        console.error("Error enviando correo:", error);
+      }
+    }
+    
+
+    return [nuevaReunion];
 
   } catch (error) {
     console.error("Error al crear reunión:", error);
-    return [null, "Error interno del servidor"];
+    return ["Error interno del servidor"];
   }
 }
 
 export async function getAllReunionesService() {
   try {
     const reunionRepo = AppDataSource.getRepository("Reunion");
-    const reuniones = await reunionRepo.find({ order: { fechaInicio: "ASC" } });
-    return [reuniones, null];
+
+    const reuniones = await reunionRepo.find({
+      relations: { acta: true }, 
+      order: { fechaInicio: "ASC" }
+    });
+
+    return [reuniones];
   } catch (error) {
     console.error("Error al obtener reuniones:", error);
-    return [null, "Error interno del servidor"];
+    return [ "Error interno del servidor"];
   }
 }
+
 
 export async function cancelarReunionService(id, motivo) {
   try {
@@ -56,25 +83,54 @@ export async function cancelarReunionService(id, motivo) {
 
     const reunion = await reunionRepo.findOne({ where: { id } });
     if (!reunion) {
-      return [null, "Reunión no encontrada"];
+      return [ "Reunión no encontrada"];
     }
 
     if (reunion.estado === "cancelada") {
-      return [null, "La reunión ya está cancelada"];
+      return [ "La reunión ya está cancelada"];
+    }
+
+    if (reunion.estado === "realizada") {
+      return [ "No puedes cancelar una reunión que ya fue realizada"];
+    }
+
+    if (!motivo || motivo == "") {
+      motivo = "Sin motivo especificado";
     }
 
     reunion.estado = "cancelada";
-    reunion.motivoCancelacion = motivo || "Sin motivo especificado";
+    reunion.motivoCancelacion = motivo
 
     await reunionRepo.save(reunion);
 
-    return [reunion, null];
+
+    const usuarioRepo = AppDataSource.getRepository("Usuario");
+
+    const usuarios = await usuarioRepo.find();
+
+    for (const usuario of usuarios) {
+      try {
+        await transporter.sendMail({
+          from: '"Junta de Vecinos" <gabriels.guzmanc@gmail.com>', 
+          to: usuario.email,  
+          subject: `Reunión Cancelada Fecha: ${reunion.fechaInicio}`,
+          text: `Estimados vecinos,\n\nSe ha cancelado la reunion para la fecha ${reunion.fechaInicio} con motivo ${reunion.asunto} \n\n Por razones de: ${motivo.motivo} \n\nSaludos cordiales,\nJunta de Vecinos Parque Ecuador`,
+        });
+        console.log("Correo enviado correctamente");
+      } catch (error) {
+        console.error("Error enviando correo:", error);
+      }
+    }
+
+
+    return [reunion];
 
   } catch (error) {
     console.error("Error al cancelar reunión:", error);
-    return [null, "Error interno del servidor"];
+    return [ "Error interno del servidor"];
   }
 }
+
 
 export async function deleteReunionService(id) {
   try {
@@ -82,13 +138,35 @@ export async function deleteReunionService(id) {
 
     const reunion = await reunionRepo.findOne({ where: { id } });
     if (!reunion) {
-      return [null, "Reunión no encontrada"];
+      return [ "Reunión no encontrada"];
+    }
+
+    if (reunion.estado === "realizada") {
+      return ["No puedes eliminar una reunión que ya fue realizada"];
     }
 
     await reunionRepo.remove(reunion);
     return [reunion, null];
   } catch (error) {
     console.error("Error al eliminar reunión:", error);
-    return [null, "Error interno del servidor"];
+    return [ "Error interno del servidor"];
   }
+}
+
+export async function finalizarReunionService(id) {
+  const repo = AppDataSource.getRepository("Reunion");
+  const reunion = await repo.findOne({ where: { id } });
+
+  if (!reunion)
+    return ["Reunión no encontrada"];
+
+  if (reunion.estado !== "programada")
+    return ["Solo puedes finalizar reuniones programadas"];
+
+  reunion.estado = "realizada";
+  reunion.fechaActualizacion = new Date();
+
+  await repo.save(reunion);
+
+  return [reunion];
 }
